+++ src/main/java/com/github/theredbrain/bettercombatextension/mixin/bettercombat/PlayerAttackHelperMixin.java
@@ -55,7 +55,7 @@ public abstract class PlayerAttackHelperMixin {
 				ItemStack mainHandStack = player.getMainHandStack();
 				ItemStack offHandStack = player.getOffHandStack();
 				setAttributesForOffHandAttack(player, true);
-				if (BetterCombatExtension.SERVER_CONFIG.enable_experimental_swap_hand_attributes_algorithm.get()) {
+				if (BetterCombatExtension.shouldAlternativeHandSwapAlgorithmBeEnabled()) {
 					player.equipStack(EquipmentSlot.MAINHAND, offHandStack);
 					player.equipStack(EquipmentSlot.OFFHAND, offHandStack);
 				} else {
@@ -63,7 +63,7 @@ public abstract class PlayerAttackHelperMixin {
 					inventory.offHand.set(0, offHandStack);
 				}
 				runnable.run();
-				if (BetterCombatExtension.SERVER_CONFIG.enable_experimental_swap_hand_attributes_algorithm.get()) {
+				if (BetterCombatExtension.shouldAlternativeHandSwapAlgorithmBeEnabled()) {
 					player.equipStack(EquipmentSlot.MAINHAND, mainHandStack);
 					player.equipStack(EquipmentSlot.OFFHAND, offHandStack);
 				} else {